+++ src/Sentry/SentryOptions.cs
@@ -41,6 +41,10 @@ public class SentryOptions
     /// </remarks>
     internal IScopeStackContainer? ScopeStackContainer { get; set; }
 
+    private Lazy<string?> LazyInstallationId => new(()
+        => new InstallationIdHelper(this).TryGetInstallationId());
+    internal string? InstallationId => LazyInstallationId.Value;
+
 #if __MOBILE__
     private bool _isGlobalModeEnabled = true;
     /// <summary>
@@ -1756,63 +1760,5 @@ public class SentryOptions
     {
         // In the future, this will most likely contain process ID
         return TryGetDsnSpecificCacheDirectoryPath();
-    }
-
-    private EventScrubber? _eventScrubber;
-    internal EventScrubber? EventScrubber
-    {
-        get
-        {
-            if (_denyList.Count == 0)
-            {
-                return null;
-            }
-            _eventScrubber ??= new EventScrubber(_denyList); // Instantiate and cache the scrubber only if needed
-            return _eventScrubber;
-        }
-        set => _eventScrubber = value;
-    }
-
-    private List<string> _denyList = EventScrubber.DefaultDenylist.ToList();
-
-    /// <summary>
-    /// By default, sensitive data is scrubbed from events that get sent to Sentry. This option allows you to disable
-    /// this default behavior. Alternatively you can use <see cref="ScrubData(IEnumerable{string})"/> or
-    /// <see cref="ScrubData(Action{IList{string}})"/> to configure scrubbing behaviour.
-    /// </summary>
-    public void DisableScrubbing()
-    {
-        _denyList = [];
-    }
-
-    /// <summary>
-    /// <para>
-    /// By default, sensitive data is scrubbed from events that get sent to Sentry using a "deny list" that contains
-    /// the names of various keys that are commonly used to store sensitive data. If items from the deny list are found
-    /// in Cookies, Headers, StackFrames etc. then the value for these items will be scrubbed and replaced with `*****`.
-    /// </para>
-    /// <para>
-    /// This method allows you to override the default deny list that Sentry uses with your own list of strings used to
-    /// identify potentially sensitive data.
-    /// </para>
-    /// </summary>
-    public void ScrubData(IEnumerable<string> denyList)
-    {
-        _denyList = denyList.ToList();
-    }
-
-    /// <summary>
-    /// <para>
-    /// By default, sensitive data is scrubbed from events that get sent to Sentry using a "deny list" that contains
-    /// the names of various keys that are commonly used to store sensitive data. If items from the deny list are found
-    /// in Cookies, Headers, StackFrames etc. then the value for these items will be scrubbed and replaced with `*****`.
-    /// </para>
-    /// <para>
-    /// This method allows you to modify the list of strings used to identify potentially sensitive data.
-    /// </para>
-    /// </summary>
-    public void ScrubData(Action<IList<string>> configureDenyList)
-    {
-        configureDenyList.Invoke(_denyList);
     }
 }